requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
numpy>=1.24.0
matplotlib>=3.7.0
wandb>=0.15.0
gymnasium>=0.28.0
stable-baselines3>=2.0.0
Pillow>=10.0.0
tqdm>=4.65.0
scikit-learn>=1.3.0
